feature: rename flink kubernetes template detail to update
