+++ scaleph-ui-react/src/pages/Project/Workspace/Kubernetes/Template/Update/Advanced/AdvancedFaultTolerance.tsx
