+++ scaleph-ui-react/src/pages/Project/Workspace/Kubernetes/Template/Update/Advanced/AdvancedHighAvailability.tsx
