+++ scaleph-ui-react/src/pages/Project/Workspace/Kubernetes/Template/Update/Advanced/AdvancedRestart.tsx
