+++ scaleph-ui-react/src/pages/Project/Workspace/Kubernetes/Template/Update/YAML/TemplateEditor.tsx
@@ -23,7 +23,7 @@ const TemplateEditor: React.FC<Props<WsFlinkKubernetesTemplate>> = (props: any) 
       height="600px"
       language="yaml"
       theme="vs-white"
-      value={props.flinkKubernetesTemplateDetail.templateYaml}
+      value={props.flinkKubernetesTemplateUpdate.templateYaml}
       options={{
         selectOnLineNumbers: true,
         readOnly: true,
@@ -35,5 +35,5 @@ const TemplateEditor: React.FC<Props<WsFlinkKubernetesTemplate>> = (props: any) 
     />
   );
 }
-const mapModelToProps = ({flinkKubernetesTemplateDetail}: any) => ({flinkKubernetesTemplateDetail})
+const mapModelToProps = ({flinkKubernetesTemplateUpdate}: any) => ({flinkKubernetesTemplateUpdate})
 export default connect(mapModelToProps)(TemplateEditor);
