+++ scaleph-ui-react/src/pages/Project/Workspace/Kubernetes/Template/Update/index.tsx
@@ -2,12 +2,12 @@ import {connect, history, useIntl, useLocation} from "umi";
 import React from "react";
 import {Button, Tabs} from "antd";
 import {FooterToolbar, PageContainer} from "@ant-design/pro-components";
-import DeploymentTemplateAdvanced from "@/pages/Project/Workspace/Kubernetes/Template/Detail/Advanced";
-import DeploymentTemplateYAML from "@/pages/Project/Workspace/Kubernetes/Template/Detail/YAML";
+import DeploymentTemplateAdvanced from "@/pages/Project/Workspace/Kubernetes/Template/Update/Advanced";
+import FlinkKubernetesDeploymentTemplateYAML from "@/pages/Project/Workspace/Kubernetes/Template/Update/YAML";
 import {WsFlinkKubernetesTemplate} from "@/services/project/typings";
 import {WsFlinkKubernetesTemplateService} from "@/services/project/WsFlinkKubernetesTemplateService";
 
-const FlinkKubernetesDeploymentTemplateDetailWeb: React.FC = (props: any) => {
+const FlinkKubernetesDeploymentTemplateUpdateWeb: React.FC = (props: any) => {
   const intl = useIntl();
   const data = useLocation().state as WsFlinkKubernetesTemplate
 
@@ -16,8 +16,8 @@ const FlinkKubernetesDeploymentTemplateDetailWeb: React.FC = (props: any) => {
   }
 
   const onSubmit = () => {
-    if (props.flinkKubernetesTemplateDetail.template) {
-      WsFlinkKubernetesTemplateService.updateTemplate(props.flinkKubernetesTemplateDetail.template).then((response) => {
+    if (props.flinkKubernetesTemplateUpdate.template) {
+      WsFlinkKubernetesTemplateService.updateTemplate(props.flinkKubernetesTemplateUpdate.template).then((response) => {
         if (response.success) {
           history.back()
         }
@@ -27,10 +27,10 @@ const FlinkKubernetesDeploymentTemplateDetailWeb: React.FC = (props: any) => {
 
   const items = [
     {label: 'Advanced', key: 'advanced', children: <DeploymentTemplateAdvanced data={data}/>},
-    {label: 'YAML', key: 'yaml', children: <DeploymentTemplateYAML data={data}/>},
+    {label: 'YAML', key: 'yaml', children: <FlinkKubernetesDeploymentTemplateYAML data={data}/>},
   ]
   return (
-    <PageContainer title={intl.formatMessage({id: 'pages.project.flink.kubernetes.template.detail'})}>
+    <PageContainer title={intl.formatMessage({id: 'pages.project.flink.kubernetes.template.update'})}>
       <Tabs items={items}/>
       <FooterToolbar>
         <Button onClick={onCancel}>{intl.formatMessage({id: 'app.common.operate.cancel.label'})}</Button>
@@ -40,5 +40,5 @@ const FlinkKubernetesDeploymentTemplateDetailWeb: React.FC = (props: any) => {
   );
 }
 
-const mapModelToProps = ({flinkKubernetesTemplateDetail}: any) => ({flinkKubernetesTemplateDetail})
-export default connect(mapModelToProps)(FlinkKubernetesDeploymentTemplateDetailWeb);
+const mapModelToProps = ({flinkKubernetesTemplateUpdate}: any) => ({flinkKubernetesTemplateUpdate})
+export default connect(mapModelToProps)(FlinkKubernetesDeploymentTemplateUpdateWeb);
